oci-hpc-clusternetwork_slurm_web_portal/function/requirements.txt
fdk
oci
paramiko
pyyaml
